+++ bindings/nodejs/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 1.0.7 - 2023-08-29
+
+### Fixed
+
+- Migration mismatch from `iota-rs` version;
+
 ## 1.0.6 - 2023-08-25
 
 ### Fixed